Separate fine-tuning methods from ifh.py, and some other small improvements.
